pyproject.toml
[tool.poetry]
name = "avitoproject"
version = "0.1.0"
description = ""
authors = ["Medved"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
django-rest-framework = "^0.1.0"
jwt = "^1.3.1"
simplejwt = "^2.0.1"
djangorestframework-simplejwt = "^5.2.2"
python-dateutil = "^2.8.2"
factory-boy = "^3.2.1"
faker = "^17.6.0"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"
pytest = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
